+++ packages/serverless-components/aws-cloudfront/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.47-alpha.1](https://github.com/getjerry/serverless-next.js/compare/@getjerry/aws-cloudfront@1.2.43...@getjerry/aws-cloudfront@1.2.47-alpha.1) (2022-11-23)
+
+**Note:** Version bump only for package @getjerry/aws-cloudfront
+
 ## [1.2.47-alpha.0](https://github.com/getjerry/serverless-next.js/compare/@getjerry/aws-cloudfront@1.2.43...@getjerry/aws-cloudfront@1.2.47-alpha.0) (2022-08-16)
 
 **Note:** Version bump only for package @getjerry/aws-cloudfront