+++ packages/serverless-components/aws-lambda/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.18-alpha.1](https://github.com/getjerry/serverless-next.js/compare/@getjerry/aws-lambda@1.3.14...@getjerry/aws-lambda@1.3.18-alpha.1) (2022-11-23)
+
+### Reverts
+
+- Revert "refactor: add murmurhash" ([6db95b4](https://github.com/getjerry/serverless-next.js/commit/6db95b48ffefa6670d706140774c8a1c87b84bf3))
+
 ## [1.3.18-alpha.0](https://github.com/getjerry/serverless-next.js/compare/@getjerry/aws-lambda@1.3.14...@getjerry/aws-lambda@1.3.18-alpha.0) (2022-08-16)
 
 **Note:** Version bump only for package @getjerry/aws-lambda